Extract test case parsing into TestCaseExtractor objects
